Players play in games which are full of stats
Decided to follow the path of having Players be the high level object,
and within a Player is a list of games in which they've played. For each
game, there are stats kept.

Probably going to have every stat be kept in a Statistic list and the
create the percentages/half/plus-minus/etc stats on the fly when a query
comes in. Actually, I dunno how the hell I'm going to do plus minus...